Move ThankYouPage to pages
