--- packages/component-submission/client/components/ThankYou.md
@@ -1,5 +1,0 @@
-## ThankYou Component
-
-```js
-;<ThankYou title="Test Manuscript Title" />
-```
