+++ packages/component-submission/client/index.js
@@ -1,2 +1,2 @@
 export { default as SubmissionWizard } from './pages/SubmissionWizard'
-export { default as ThankYouPage } from './pages/ThankYou'
+export { default as ThankYouPage } from './pages/ThankYouPage'
