+++ packages/component-submission/client/pages/ThankYouPage.test.js
@@ -1,13 +1,23 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import ThankYou from './ThankYou'
+import { ThankYouPage } from './ThankYouPage'
 
 describe('ThankYou', () => {
   it('has a hotjar supress attribute on the element around the title', () => {
     const originalWindow = global.window
     global.window.hj = jest.fn
 
-    const wrapper = shallow(<ThankYou title="Test Manuscript Title" />)
+    const wrapper = shallow(
+      <ThankYouPage
+        data={{
+          manuscript: {
+            meta: {
+              title: 'Hello World',
+            },
+          },
+        }}
+      />,
+    )
     expect(
       wrapper.find('[data-test-id="title"][data-hj-suppress=""]'),
     ).toHaveLength(1)
